split up tool.py into two
