+++ src/_overrides/__init__.py
@@ -1,0 +1,14 @@
+# This module contains all of our dirty lil override hacks
+#   Usually this means matplotlib overriding of Toolbars,
+#   Axes, and such.
+
+# These are HACKS. They WILL break if you update the
+#   packages. If we could do these things without doing
+#   them this way, we would have, but we can't as far
+#   as we know.
+
+# As such, there is a lesser standard of quality to the
+#   modules under this package.
+
+# We generally suggest you close your eyes and ears and
+#   wonder out this godforsaken space.
